+++ cvrp_ocl/move.h
@@ -18,14 +18,14 @@ enum MoveType{
 
 class Move {
 public:
-    Move(AntStruct *ant_, bool valid_, double gain_, int pos_n1_, int pos_n2_, MoveType type_)
+    Move(AntStruct *ant_, bool valid_, float gain_, int pos_n1_, int pos_n2_, MoveType type_)
     :ant(ant_), valid(valid_), gain(gain_), pos_n1(pos_n1_), pos_n2(pos_n2_), type(type_){}
     virtual ~Move(){}
     virtual void apply(){}
     
     AntStruct *ant;
     bool valid;
-    double   gain;
+    float   gain;
     int pos_n1;          /* 两个互换位置的node idx */
     int pos_n2;
     
@@ -34,7 +34,7 @@ public:
 
 class ExchangeMove : public Move {
 public:
-    ExchangeMove(AntStruct *ant_, bool valid_, double gain_, int pos_n1_, int pos_n2_,
+    ExchangeMove(AntStruct *ant_, bool valid_, float gain_, int pos_n1_, int pos_n2_,
                  int load_r1_, int load_r2_)
     :Move(ant_, valid_, gain_, pos_n1_, pos_n2_, EXCHANGE_MOVE), load_r1(load_r1_), load_r2(load_r2_){}
     ~ExchangeMove(){}
@@ -48,7 +48,7 @@ public:
 
 class InsertionMove : public Move {
 public:
-    InsertionMove(AntStruct *ant_, bool valid_, double gain_, int pos_n1_, int pos_n2_,
+    InsertionMove(AntStruct *ant_, bool valid_, float gain_, int pos_n1_, int pos_n2_,
                  int load_r1_, int load_r2_)
     :Move(ant_, valid_, gain_, pos_n1_, pos_n2_, INSERTION_MOVE), load_r1(load_r1_), load_r2(load_r2_){}
     ~InsertionMove(){}
@@ -61,7 +61,7 @@ public:
 
 class InversionMove : public Move {
 public:
-    InversionMove(AntStruct *ant_, bool valid_, double gain_, int pos_n1_, int pos_n2_)
+    InversionMove(AntStruct *ant_, bool valid_, float gain_, int pos_n1_, int pos_n2_)
     :Move(ant_, valid_, gain_, pos_n1_, pos_n2_, INVERSION_MOVE){}
     ~InversionMove(){}
