+++ cvrp_ocl/unix_timer.cpp
@@ -27,7 +27,7 @@
 
 static struct rusage res;
 static struct timeval tp;
-static double virtual_time, real_time;
+static float virtual_time, real_time;
 static char format_time[26];
 
 
@@ -43,19 +43,19 @@ void start_timers(void)
 */
 {
     getrusage( RUSAGE_SELF, &res );
-    virtual_time = (double) res.ru_utime.tv_sec +
-		   (double) res.ru_stime.tv_sec +
-		   (double) res.ru_utime.tv_usec / 1000000.0 +
-		   (double) res.ru_stime.tv_usec / 1000000.0;
+    virtual_time = (float) res.ru_utime.tv_sec +
+		   (float) res.ru_stime.tv_sec +
+		   (float) res.ru_utime.tv_usec / 1000000.0 +
+		   (float) res.ru_stime.tv_usec / 1000000.0;
 
     gettimeofday( &tp, NULL );
-    real_time =    (double) tp.tv_sec +
-		   (double) tp.tv_usec / 1000000.0;
+    real_time =    (float) tp.tv_sec +
+		   (float) tp.tv_usec / 1000000.0;
 }
 
 
 
-double elapsed_time(TIMER_TYPE type)
+float elapsed_time(TIMER_TYPE type)
 /*    
       FUNCTION:       return the time used in seconds (virtual or real, depending on type) 
       INPUT:          TIMER_TYPE (virtual or real time)
@@ -65,16 +65,16 @@ double elapsed_time(TIMER_TYPE type)
 {
     if (type == REAL) {
         gettimeofday( &tp, NULL );
-        return( (double) tp.tv_sec +
-		(double) tp.tv_usec / 1000000.0
+        return( (float) tp.tv_sec +
+		(float) tp.tv_usec / 1000000.0
 		- real_time );
     }
     else {
         getrusage( RUSAGE_SELF, &res );
-        return( (double) res.ru_utime.tv_sec +
-		(double) res.ru_stime.tv_sec +
-		(double) res.ru_utime.tv_usec / 1000000.0 +
-		(double) res.ru_stime.tv_usec / 1000000.0
+        return( (float) res.ru_utime.tv_sec +
+		(float) res.ru_stime.tv_sec +
+		(float) res.ru_utime.tv_usec / 1000000.0 +
+		(float) res.ru_stime.tv_usec / 1000000.0
 		- virtual_time );
     }
